+++ test/src/unit-capi-config.cc
@@ -192,6 +192,7 @@ void check_save_to_file() {
   ss << "sm.array_schema_cache_size 10000000\n";
   ss << "sm.check_coord_dups true\n";
   ss << "sm.check_coord_oob true\n";
+  ss << "sm.check_global_order true\n";
   ss << "sm.dedup_coords false\n";
   ss << "sm.enable_signal_handlers true\n";
   ss << "sm.fragment_metadata_cache_size 10000000\n";
@@ -363,6 +364,7 @@ TEST_CASE("C API: Test config iter", "[capi], [config]") {
   all_param_values["sm.dedup_coords"] = "false";
   all_param_values["sm.check_coord_dups"] = "true";
   all_param_values["sm.check_coord_oob"] = "true";
+  all_param_values["sm.check_global_order"] = "true";
   all_param_values["sm.tile_cache_size"] = "100";
   all_param_values["sm.array_schema_cache_size"] = "1000";
   all_param_values["sm.fragment_metadata_cache_size"] = "10000000";